feat: redesign in navigation bar and implement a new profile screen
