+++ app/src/main/java/com/free2party/MainViewModel.kt
@@ -10,16 +10,19 @@ import com.free2party.data.model.FriendRequestStatus
 import com.free2party.data.model.Notification
 import com.free2party.data.model.NotificationType
 import com.free2party.data.model.ThemeMode
+import com.free2party.data.repository.PlanRepository
 import com.free2party.data.repository.SettingsRepository
 import com.free2party.data.repository.SocialRepository
 import com.free2party.data.repository.UserRepository
 import com.free2party.ui.navigation.Screen
+import com.free2party.util.isPlanActive
 import dagger.hilt.android.lifecycle.HiltViewModel
 import javax.inject.Inject
 import kotlinx.coroutines.channels.Channel
 import kotlinx.coroutines.flow.MutableSharedFlow
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.asSharedFlow
+import kotlinx.coroutines.flow.asStateFlow
 import kotlinx.coroutines.flow.collectLatest
 import kotlinx.coroutines.flow.combine
 import kotlinx.coroutines.flow.first
@@ -29,13 +32,22 @@ import kotlinx.coroutines.withTimeoutOrNull
 import kotlinx.coroutines.launch
 import kotlin.time.Duration.Companion.milliseconds
 
+data class UserNavState(
+    val profilePicUrl: String = "",
+    val isUserFree: Boolean = false
+)
+
 @HiltViewModel
 class MainViewModel @Inject constructor(
     private val settingsRepository: SettingsRepository,
     private val socialRepository: SocialRepository,
-    private val userRepository: UserRepository
+    private val userRepository: UserRepository,
+    private val planRepository: PlanRepository
 ) : ViewModel() {
     private var initialHandledNotificationId: String? = null
+
+    private val _userNavState = MutableStateFlow<UserNavState?>(null)
+    val userNavState = _userNavState.asStateFlow()
 
     fun setInitialNotificationId(id: String?) {
         initialHandledNotificationId = id
@@ -80,6 +92,7 @@ class MainViewModel @Inject constructor(
         observeUserSettings()
         observeFriendRequests()
         observeNotifications()
+        observeUserNavState()
     }
 
     private fun observeThemeMode() {
@@ -263,6 +276,31 @@ class MainViewModel @Inject constructor(
         }
     }
 
+    private fun observeUserNavState() {
+        viewModelScope.launch {
+            userRepository.userIdFlow.collectLatest { uid ->
+                if (uid.isNotBlank()) {
+                    combine(
+                        userRepository.observeUser(uid),
+                        planRepository.getOwnPlans()
+                    ) { user, plans ->
+                        val isAnyPlanActiveNow = plans.any { isPlanActive(it) }
+                        val effectiveIsFree =
+                            if (user.isStatusFromPlan) isAnyPlanActiveNow else user.isFreeNow
+                        UserNavState(
+                            profilePicUrl = user.profilePicUrl,
+                            isUserFree = effectiveIsFree
+                        )
+                    }.collect { state ->
+                        _userNavState.value = state
+                    }
+                } else {
+                    _userNavState.value = null
+                }
+            }
+        }
+    }
+
     fun onNotificationClicked(notificationId: String) {
         activeSystemNotifications.update { it + notificationId }
 
@@ -279,7 +317,12 @@ class MainViewModel @Inject constructor(
                 }
                 if (notification.eventId.isNotBlank()) {
                     val scrollToComments = notification.type == NotificationType.EVENT_COMMENT
-                    _navigateToRoute.emit(Screen.EventDetails.createRoute(notification.eventId, scrollToComments))
+                    _navigateToRoute.emit(
+                        Screen.EventDetails.createRoute(
+                            notification.eventId,
+                            scrollToComments
+                        )
+                    )
                 } else {
                     _navigateToRoute.emit(Screen.Notifications.route)
                 }
