+++ app/src/main/java/com/free2party/ui/components/PlanResults.kt
@@ -19,6 +19,7 @@ import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.dp
 import com.free2party.R
 import com.free2party.data.model.FriendInfo
@@ -49,6 +50,7 @@ fun PlanResults(
         Text(
             text = stringResource(R.string.results_for, selectedDateText),
             style = MaterialTheme.typography.titleMedium,
+            fontWeight = FontWeight.Bold,
             color = MaterialTheme.colorScheme.onSurface
         )
