+++ firestore.rules
@@ -135,7 +135,11 @@ service cloud.firestore {
           request.auth.uid in get(/databases/$(database)/documents/events/$(eventId)).data.guestIds
         );
         
-        allow update, delete: if request.auth != null && resource.data.userId == request.auth.uid;
+        allow update: if request.auth != null && resource.data.userId == request.auth.uid;
+        allow delete: if request.auth != null && (
+          resource.data.userId == request.auth.uid ||
+          get(/databases/$(database)/documents/events/$(eventId)).data.hostId == request.auth.uid
+        );
       }
 
       // Photos subcollection
